+++ src/Aws/Ses/Resources/ses-2010-12-01.php
@@ -723,6 +723,9 @@ return array (
                                     'ComplaintTopic' => array(
                                         'type' => 'string',
                                     ),
+                                    'DeliveryTopic' => array(
+                                        'type' => 'string',
+                                    ),
                                     'ForwardingEnabled' => array(
                                         'type' => 'boolean',
                                     ),